Revert loop sound fix and fix it properly
Commit 529b6568 was a bit over overzealous. It caused new server
crashes. The check for soundSet to be a valid string was not the
correct answer to distinguish G_SoundIndex loopSounds and
CAS_GetBModelSound loopSounds. It seems like it can happen that
loopSounds is a CAS_GetBModelSound index, but soundSet is NULL.

So instead we address the original server crash on ns_hideout directly.
The problem there is that turrets, which usually are not ET_MOVER are
moved using the Q3_Lerp* functions, which make them ET_MOVER. So we
simply omit the sound now if they were made ET_MOVERs.