Fix crash due to wrong clearing of list
Frankly i'm not quite sure what that implementation is supposed to be.
Clearly the return value of erase should be use to set the next l
instead of the increment. Maybe that code just happened to work with
their standard library implementation? Anyways, don't see a reason why
one wouldn't just call .clear().