Add rule restrictions for sv_fps cvar
Speedrun rules allow only the values 20 and 125 for sv_fps. So those
restrictions have to be checked. Additionally, it is not allowed to
switch between 20 and 125 quickly at any time. As compromise, we allow
changing the value for each map. To implement that, we abuse the
"latching" of cvars that usually applies changes to such latched cvars
on vid_restart (mainly graphics settings). With some hacks we make
sv_fps "latch" until the next map and then apply the new value.