Allow forcing RGBA8 for internal texture format
Using the same internal format as the source format greatly reduces the
time it takes to load the textures with OpenGL. Adding a cvar to force
internal texture format to be RGBA8, which is the input format for all
textures. May still disable in case there are systems that have better
performance with the original behaviour of using all kinds of different
internal formats.